Made the dropdown searchable for building and room
Currently, the dropdown for rooms only display the BBB rooms, but it will be fixed in the future to show the corresponding rooms for the current building shown on the building dropdown. Also, it will be fixed that the buildingInfoStore will only fetch the building data once when MainActivity is started.