Path generator written but untested
Wrote out the A* algorithm and some associated necessary changes to Node structure, and some helper functions. I haven't tested it yet, and there's still a couple weird Kotlin specific (or maybe Android Studio specific) errors that I need to figure out. However this gives a good basic structure for others to use as a black box when implementing things that rely on path generation.